+++ src/main/java/org/softserve/Config.java
@@ -1,5 +1,9 @@
 package org.softserve;
 
 public class Config {
-    public static final String BROWSER = "";
+    public static final String BROWSER = "Chrome";
+
+    public static final String LOGIN = "ivankoprogamer@gmail.com";
+    public static final String PASSWORD = "qwe123QWE";
+
 }
