+++ src/main/java/org/softserve/pages/logout/LogoutPage.java
@@ -3,7 +3,8 @@ package org.softserve.pages.logout;
 import org.openqa.selenium.WebDriver;
 import org.softserve.pages.base.BasePage;
 
-import static org.softserve.constants.Constant.Links.LOGOUT_LINK;
+
+import static org.softserve.constants.Constants.Links.LOGOUT_LINK;
 
 public class LogoutPage extends BasePage {
     public LogoutPage(WebDriver driver) {
